requirements.txt
auto_mix_prep==0.2.0
azure-storage-blob==12.26.0
dotenv==0.9.9
fastapi==0.116.1
pandas==2.3.2
pydantic==2.11.7
scipy==1.16.1
statsmodels==0.14.5
uvicorn==0.30.3
